pyproject.toml
[tool.poetry]
name = "atom"
version = "1.0"
description = "An end-to-end library for acoustic tomography in the atmosphere."
authors = ["Nicholas Hamilton <nicholas.hamilton@nrel.gov>"]
license = "LICENSE.md"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
hydra-core = "1.3.2"
matplotlib = "3.9.4"
numpy = "2.2.0"
pandas = "2.2.3"
scipy = "1.14.1"
setuptools = "75.1.0"
xarray = "2023.6.0"
netcdf4 = "^1.7.2"
jupyterlab = "^4.3.3"
jupyter-book = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
